serverless/requirement2.txt
anytree==2.8.0
boto3==1.20.10
botocore==1.23.10
certifi==2024.7.4
charset-normalizer==2.0.7
et-xmlfile==1.1.0
excel2json-3==0.1.6
idna==3.3
jmespath==0.10.0
numpy==1.22.2
openpyxl==3.0.9
pandas==1.1.5
pydash==6.0.0
pymongo==3.10.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.31.0
s3transfer==0.5.0
six==1.16.0
termcolor==1.1.0
urllib3==1.26.18
xlrd==1.2.0
XlsxWriter==3.0.2
xlutils==2.0.0
xlwt==1.3.0
watchtower==3.0.1
psutil==5.9.8
fastapi
requests